sheets_sync/requirements.txt
requests==2.31.0
gspread==6.0.0
google-auth==2.27.0
pytz==2024.1
apscheduler==3.10.4
python-dotenv==1.0.1
